Fix #18 by disabling PW-leak-prompt (#23)
The WebDriver instance for Chrome showed an Password-Breach-Prompt a few miliseconds after logging in. The prompt made it impossible to click on elements of the webpage, which caused tests to fail. We disabled the prompt.

We did 10 runs of all tests.
Result: All passed, no reruns.
Conclusion: This might have been the only reason for flaky tests. It might no longer be necessary to mark the inventory_tests as flaky. But we keep it for now and decide later.